+++ app/src/main/java/com/architect/coders/pokedex/data/database/PokemonDatabase.kt
@@ -1,0 +1,13 @@
+package com.architect.coders.pokedex.data.database
+
+import androidx.room.Database
+import androidx.room.RoomDatabase
+
+@Database(
+    entities = [PokemonL::class, TypeL::class, StatL::class, CollectionL::class],
+    version = 1,
+    exportSchema = false
+)
+abstract class PokemonDatabase: RoomDatabase() {
+    abstract fun pokemonDao(): PokemonDAO
+}
