+++ app/src/main/java/com/architect/coders/pokedex/data/database/PokemonDetailL.kt
@@ -1,4 +1,4 @@
-package com.architect.coders.pokedex.database
+package com.architect.coders.pokedex.data.database
 
 import androidx.room.Embedded
 import androidx.room.Relation
