+++ app/src/main/java/com/architect/coders/pokedex/data/network/StatR.kt
@@ -1,5 +1,4 @@
-package com.architect.coders.pokedex.network
-
+package com.architect.coders.pokedex.data.network
 
 import com.google.gson.annotations.SerializedName
