+++ app/src/main/java/com/architect/coders/pokedex/data/network/TypeX.kt
@@ -1,9 +1,8 @@
-package com.architect.coders.pokedex.network
-
+package com.architect.coders.pokedex.data.network
 
 import com.google.gson.annotations.SerializedName
 
-data class Species(
+data class TypeX(
     @SerializedName("name")
     val name: String,
     @SerializedName("url")
