+++ app/src/main/java/com/architect/coders/pokedex/model/GalleryItem.kt
@@ -1,5 +1,5 @@
 package com.architect.coders.pokedex.model
 
-import com.architect.coders.pokedex.common.PokeCollec
+import com.architect.coders.pokedex.ui.common.PokeCollec
 
 data class GalleryItem(val type: PokeCollec, val photos: MutableList<String>)
