+++ app/src/main/java/com/architect/coders/pokedex/ui/common/PictureRequester.kt
@@ -1,4 +1,4 @@
-package com.architect.coders.pokedex.model
+package com.architect.coders.pokedex.ui.common
 
 import android.net.Uri
 import androidx.activity.result.contract.ActivityResultContracts
@@ -6,7 +6,7 @@ import androidx.fragment.app.Fragment
 import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlin.coroutines.resume
 
-class DispatchPicture(fragment: Fragment) {
+class PictureRequester(fragment: Fragment) {
 
     private var onRequest: (Boolean) -> Unit = {}
     private val launcher = fragment.registerForActivityResult(ActivityResultContracts.TakePicture()) { success ->
