+++ app/src/main/res/layout/fragment_collection.xml
@@ -7,7 +7,7 @@
     tools:context=".ui.collection.CollectionFragment"
     android:background="@android:color/darker_gray">
 
-    <com.architect.coders.pokedex.util.ImageZoom
+    <com.architect.coders.pokedex.ui.collection.ImageZoom
         android:id="@+id/imvCollection"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
